requirements.txt
gradio
numpy
pandas
scipy
matplotlib
tensorflow
pyyaml
